Trocando tipo de Usuario e Premio de Integer para Long

+++ src/main/java/br/com/topsorteio/dtos/MarcaRegisterRequestDTO.java
@@ -2,4 +2,4 @@ package br.com.topsorteio.dtos;
 
 import org.springframework.web.multipart.MultipartFile;
 
-public record MarcaRegisterRequestDTO(String nome, String titulo, MultipartFile logo, MultipartFile banner, int ordemExibicao, Integer criadoPor) {}
+public record MarcaRegisterRequestDTO(String nome, String titulo, MultipartFile logo, MultipartFile banner, int ordemExibicao, Long criadoPor) {}
