+++ src/main/java/br/com/topsorteio/repositories/iUserRepository.java
@@ -13,7 +13,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface iUserRepository extends JpaRepository<UserModel, Integer> {
+public interface iUserRepository extends JpaRepository<UserModel, Long> {
     Optional<UserModel> findByEmail(String email);
 
 }
